Progress on player-ball collision
Horizontal collision ok, vertical collisions still need more work.
The ball.pos + ball.radius logic is incomplete, need to multiply the
radius with the current speed of the ball to know whether or not it
needs to be added or substracted from the coordinates.

Add: checkHorizontalCollisions() and checkVerticalCollisions() to know
when to switch the current speed of the ball.