Added protection against multiple clients with the same id connecting
consumer now notifies the client that they are already in a game and
must either leave it or finish it before starting another one.
consumer checks in accept() for a matching player_id and goes into the
branch if the game_room.missing_player id is not the same as the one
coming in.

changed missing_player variable in game_room to hold the id of the
dropped out player instead of an arbitrary value. Added
get_missing_player_id() getter to be called by consumer when checking in
accept()